# Optimize `cbrtf` by removing checks for non-normal floats