@shssoichiro 2022-11-15
Nice gains